fastapi_rbac/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
Authlib==1.4.1
black==25.1.0
casbin==1.38.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.18
dnspython==2.7.0
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jwt==1.3.1
limits==4.0.1
motor==3.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11
python-dotenv==1.0.1
redis==5.2.1
setuptools==75.8.0
simpleeval==1.0.3
slowapi==0.1.9
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
